requirements.txt
# Core dependencies
aiogram==3.12.0
mistralai==0.4.2

# Video processing
yt-dlp==2024.8.6
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.5
